Added 3 missing terran upgrades: * Fact/tech lab:  RESEARCH_RAPIDFIRELAUNCHERS * Fact/tech lab:  RESEARCH_SMARTSERVOS * Starport/tech lab:  RESEARCH_ENHANCEDMUNITIONS
Added both the ABILITY_ID values as well as the UPGRADE_ID values.
NOTE:  RESEARCH_MAGFIELDLAUNCHERS is not a currently available upgrade (it is from a previous version), but is in the list with the same ID as RESEARCH_SMARTSERVOS.  I left this to avoid a breaking change and will open a separate issue for it.